requirements.txt
appdirs==1.4.4
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
distlib==0.3.1
filelock==3.0.12
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
pyTelegramBotAPI==3.7.6
requests==2.25.1
six==1.15.0
toml==0.10.2
urllib3==1.26.3
virtualenv==20.4.0
wrapt==1.12.1
wget==3.2
python-telegram-bot==13.1
Pillow==8.1.0
torch @ https://download.pytorch.org/whl/cpu/torch-1.7.1%2Bcpu-cp37-cp37m-linux_x86_64.whl
torchvision @ https://download.pytorch.org/whl/cpu/torchvision-0.8.2%2Bcpu-cp37-cp37m-linux_x86_64.whl